+++ ratman/src/daemon/mod.rs
@@ -3,6 +3,7 @@
 pub mod config;
 mod parse;
 mod peers;
+pub mod pidfile;
 mod state;
 mod transform;